+++ packages/bridge-controller/src/bridge-controller.ts
@@ -351,6 +351,7 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
     featureId: FeatureId | null = null,
   ): Promise<(QuoteResponse & L1GasFees & NonEvmFees)[]> => {
     const bridgeFeatureFlags = getBridgeFeatureFlags(this.messenger);
+    const jwt = await this.#getJwt();
     // If featureId is specified, retrieve the quoteRequestOverrides for that featureId
     const quoteRequestOverrides = featureId
       ? bridgeFeatureFlags.quoteRequestOverrides?.[featureId]
@@ -364,6 +365,7 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
         : { ...quoteRequest, resetApproval },
       abortSignal,
       this.#clientId,
+      jwt,
       this.#fetchFn,
       this.#config.customBridgeApiBaseUrl ?? BRIDGE_PROD_API_BASE_URL,
       featureId,
@@ -625,6 +627,8 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
       state.quotesLoadingStatus = RequestStatus.LOADING;
     });
 
+    const jwt = await this.#getJwt();
+
     try {
       await this.#trace(
         {
@@ -653,6 +657,7 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
           if (shouldStream) {
             await this.#handleQuoteStreaming(
               updatedQuoteRequest,
+              jwt,
               selectedAccount,
             );
             return;
@@ -739,6 +744,7 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
 
   readonly #handleQuoteStreaming = async (
     updatedQuoteRequest: GenericQuoteRequest,
+    jwt?: string,
     selectedAccount?: InternalAccount,
   ) => {
     /**
@@ -757,6 +763,7 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
       updatedQuoteRequest,
       this.#abortController?.signal,
       this.#clientId,
+      jwt,
       this.#config.customBridgeApiBaseUrl ?? BRIDGE_PROD_API_BASE_URL,
       {
         onValidationFailure: this.#trackResponseValidationFailures,
@@ -863,6 +870,18 @@ export class BridgeController extends StaticIntervalPollingController<BridgePoll
     );
     return networkClient;
   }
+
+  readonly #getJwt = async (): Promise<string | undefined> => {
+    try {
+      const token = await this.messenger.call(
+        'AuthenticationController:getBearerToken',
+      );
+      return token;
+    } catch (error) {
+      console.error('Error getting JWT token for bridge-api request', error);
+      return undefined;
+    }
+  };
 
   readonly #getRequestMetadata = (): Omit<
     RequestMetadata,