+++ packages/bridge-controller/src/utils/fetch.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/explicit-function-return-type */
 import { StructError } from '@metamask/superstruct';
 import type { CaipAssetType, CaipChainId, Hex } from '@metamask/utils';
 
@@ -18,16 +19,28 @@ import type {
   BridgeAsset,
 } from '../types';
 
-export const getClientHeaders = (clientId: string, clientVersion?: string) => ({
+export const getClientHeaders = ({
+  clientId,
+  clientVersion,
+  jwt,
+}: {
+  clientId: string;
+  clientVersion?: string;
+  jwt?: string;
+}) => ({
   'X-Client-Id': clientId,
+  ...(jwt ? { Authorization: `Bearer ${jwt}` } : {}),
   ...(clientVersion ? { 'Client-Version': clientVersion } : {}),
 });
 
 /**
  * Returns a list of enabled (unblocked) tokens
  *
+ * @deprecated Use the popular and search bridge-api endpoints instead
+ *
  * @param chainId - The chain ID to fetch tokens for
  * @param clientId - The client ID for metrics
+ * @param jwt - The JWT token for authentication
  * @param fetchFn - The fetch function to use
  * @param bridgeApiBaseUrl - The base URL for the bridge API
  * @param clientVersion - The client version for metrics (optional)
@@ -36,18 +49,18 @@ export const getClientHeaders = (clientId: string, clientVersion?: string) => ({
 export async function fetchBridgeTokens(
   chainId: Hex | CaipChainId,
   clientId: string,
+  jwt: string | undefined,
   fetchFn: FetchFunction,
   bridgeApiBaseUrl: string,
   clientVersion?: string,
 ): Promise<Record<string, BridgeAsset>> {
-  // TODO make token api v2 call
   const url = `${bridgeApiBaseUrl}/getTokens?chainId=${formatChainIdToDec(chainId)}`;
 
   // TODO we will need to cache these. In Extension fetchWithCache is used. This is due to the following:
   // If we allow selecting dest networks which the user has not imported,
   // note that the Assets controller won't be able to provide tokens. In extension we fetch+cache the token list from bridge-api to handle this
   const tokens = await fetchFn(url, {
-    headers: getClientHeaders(clientId, clientVersion),
+    headers: getClientHeaders({ clientId, clientVersion, jwt }),
   });
 
   const transformedTokens: Record<string, BridgeAsset> = {};
@@ -107,6 +120,7 @@ const formatQueryParams = (request: GenericQuoteRequest): URLSearchParams => {
  * @param request - The quote request
  * @param signal - The abort signal
  * @param clientId - The client ID for metrics
+ * @param jwt - The JWT token for authentication
  * @param fetchFn - The fetch function to use
  * @param bridgeApiBaseUrl - The base URL for the bridge API
  * @param featureId - The feature ID to append to each quote
@@ -117,6 +131,7 @@ export async function fetchBridgeQuotes(
   request: GenericQuoteRequest,
   signal: AbortSignal | null,
   clientId: string,
+  jwt: string | undefined,
   fetchFn: FetchFunction,
   bridgeApiBaseUrl: string,
   featureId: FeatureId | null,
@@ -129,7 +144,7 @@ export async function fetchBridgeQuotes(
 
   const url = `${bridgeApiBaseUrl}/getQuote?${queryParams}`;
   const quotes: unknown[] = await fetchFn(url, {
-    headers: getClientHeaders(clientId, clientVersion),
+    headers: getClientHeaders({ clientId, clientVersion, jwt }),
     signal,
   });
 
@@ -142,10 +157,10 @@ export async function fetchBridgeQuotes(
         if (error instanceof StructError) {
           error.failures().forEach(({ branch, path }) => {
             const aggregatorId =
-              branch?.[0]?.quote?.bridgeId ||
-              branch?.[0]?.quote?.bridges?.[0] ||
-              (quoteResponse as QuoteResponse)?.quote?.bridgeId ||
-              (quoteResponse as QuoteResponse)?.quote?.bridges?.[0] ||
+              branch?.[0]?.quote?.bridgeId ??
+              branch?.[0]?.quote?.bridges?.[0] ??
+              (quoteResponse as QuoteResponse)?.quote?.bridgeId ??
+              (quoteResponse as QuoteResponse)?.quote?.bridges?.[0] ??
               'unknown';
             const pathString = path?.join('.') || 'unknown';
             uniqueValidationFailures.add([aggregatorId, pathString].join('|'));
@@ -200,7 +215,7 @@ const fetchAssetPricesForCurrency = async (request: {
   });
   const url = `https://price.api.cx.metamask.io/v3/spot-prices?${queryParams}`;
   const priceApiResponse = (await fetchFn(url, {
-    headers: getClientHeaders(clientId, clientVersion),
+    headers: getClientHeaders({ clientId, clientVersion }),
     signal,
   })) as Record<CaipAssetType, { [currency: string]: number }>;
   if (!priceApiResponse || typeof priceApiResponse !== 'object') {
@@ -274,6 +289,7 @@ export const fetchAssetPrices = async (
  * @param request - The quote request
  * @param signal - The abort signal
  * @param clientId - The client ID for metrics
+ * @param jwt - The JWT token for authentication
  * @param bridgeApiBaseUrl - The base URL for the bridge API
  * @param serverEventHandlers - The server event handlers
  * @param serverEventHandlers.onValidationFailure - The function to handle validation failures
@@ -287,6 +303,7 @@ export async function fetchBridgeQuoteStream(
   request: GenericQuoteRequest,
   signal: AbortSignal | undefined,
   clientId: string,
+  jwt: string | undefined,
   bridgeApiBaseUrl: string,
   serverEventHandlers: {
     onClose: () => void | Promise<void>;
@@ -320,11 +337,11 @@ export async function fetchBridgeQuoteStream(
       if (error instanceof StructError) {
         error.failures().forEach(({ branch, path }) => {
           const aggregatorId =
-            branch?.[0]?.quote?.bridgeId ||
-            branch?.[0]?.quote?.bridges?.[0] ||
-            (quoteResponse as QuoteResponse)?.quote?.bridgeId ||
-            (quoteResponse as QuoteResponse)?.quote?.bridges?.[0] ||
-            'unknown';
+            branch?.[0]?.quote?.bridgeId ??
+            branch?.[0]?.quote?.bridges?.[0] ??
+            (quoteResponse as QuoteResponse)?.quote?.bridgeId ??
+            ((quoteResponse as QuoteResponse)?.quote?.bridges?.[0] ||
+              ('unknown' as string));
           const pathString = path?.join('.') || 'unknown';
           uniqueValidationFailures.add([aggregatorId, pathString].join('|'));
         });
@@ -343,14 +360,14 @@ export async function fetchBridgeQuoteStream(
   const urlStream = `${bridgeApiBaseUrl}/getQuoteStream?${queryParams}`;
   await fetchServerEvents(urlStream, {
     headers: {
-      ...getClientHeaders(clientId, clientVersion),
+      ...getClientHeaders({ clientId, clientVersion, jwt }),
       'Content-Type': 'text/event-stream',
     },
     signal,
     onMessage,
-    onError: (e) => {
+    onError: (error) => {
       // Rethrow error to prevent silent fetch failures
-      throw e;
+      throw error;
     },
     onClose: async () => {
       await serverEventHandlers.onClose();