+++ packages/bridge-status-controller/src/bridge-status-controller.ts
@@ -774,6 +774,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
           historyItem.quote.intent?.protocol ?? '',
           srcChainId.toString(),
           this.#clientId,
+          await this.#getJwt(),
         );
 
         intentOrderStatus = intentOrder.status;
@@ -801,6 +802,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
         const response = await fetchBridgeTxStatus(
           statusRequest,
           this.#clientId,
+          await this.#getJwt(),
           this.#fetchFn,
           this.#config.customBridgeApiBaseUrl,
         );
@@ -889,6 +891,18 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
     } catch (error) {
       console.warn('Failed to fetch bridge tx status', error);
       this.#handleFetchFailure(bridgeTxMetaId);
+    }
+  };
+
+  readonly #getJwt = async (): Promise<string | undefined> => {
+    try {
+      const token = await this.messenger.call(
+        'AuthenticationController:getBearerToken',
+      );
+      return token;
+    } catch (error) {
+      console.error('Error getting JWT token for bridge-api request', error);
+      return undefined;
     }
   };
 
@@ -1663,6 +1677,7 @@ export class BridgeStatusController extends StaticIntervalPollingController<Brid
       const intentOrder = await intentApi.submitIntent(
         submissionParams,
         this.#clientId,
+        await this.#getJwt(),
       );
 
       const orderUid = intentOrder.id;